qrc_model/requirements.txt
kafka-python==2.0.2
numpy==1.24.3
pennylane==0.33.1